Speedup new in the case where the value is already interned
This will slow down the case where value is not already interned as it
requires second hash lookup on the map to acquire the write lock if not
present.

Added additional test to check refcount bound during test. This showed
that an attempted optimization using into_key() to save a clone() for
the Entry::Occupied case in the write lock acquiring path was incorrect.